test: split test suite to extract browser and node specific tests in their own folders
