+++ test/common/index.test.js
@@ -1,5 +1,5 @@
-import * as src from '../src/index';
-import * as dist from '../dist/index';
+import * as src from '../../src';
+import * as dist from '../../dist';
 
 const regex1 = /foo/;
 const regex2 = /foo/g;
@@ -122,10 +122,6 @@ const tests = ({ stringify, parse }) => {
     const stringifiedSpaced = stringify(data, { space: 2 });
 
     expect(stringifiedSpaced).toMatchSnapshot();
-  });
-
-  test('stringify the global object', () => {
-    expect(() => stringify(global, { maxDepth: 10000 })).not.toThrow();
   });
 
   test('check duplicate value', () => {
@@ -332,48 +328,6 @@ const tests = ({ stringify, parse }) => {
     const parsed = parse(stringified);
 
     expect(parsed['foo.b'].constructor.name).toEqual('Foo');
-  });
-
-  test('HTML Event', () => {
-    const event = new MouseEvent('click', { bubbles: true, composed: true, cancelable: true });
-
-    const stringified = stringify(event);
-
-    const parsed = parse(stringified);
-
-    expect(parsed).toMatchObject({
-      isTrusted: expect.any(Boolean),
-      type: 'click',
-      bubbles: true,
-      composed: true,
-      cancelable: true,
-      returnValue: expect.any(Boolean),
-      timeStamp: expect.any(Number),
-    });
-  });
-
-  test('HTML Custom Event', () => {
-    const event = new CustomEvent('custom:click', {
-      bubbles: true,
-      composed: true,
-      cancelable: true,
-      detail: { aKey: 'a Value' },
-    });
-
-    const stringified = stringify(event, { allowClass: true });
-
-    const parsed = parse(stringified);
-
-    expect(parsed).toMatchObject({
-      isTrusted: expect.any(Boolean),
-      type: 'custom:click',
-      bubbles: true,
-      composed: true,
-      cancelable: true,
-      returnValue: expect.any(Boolean),
-      timeStamp: expect.any(Number),
-      detail: { aKey: 'a Value' },
-    });
   });
 };
