+++ test/node/index.test.js
@@ -1,0 +1,16 @@
+import * as src from '../../src';
+import * as dist from '../../dist';
+
+const tests = ({ stringify }) => {
+  test('stringify the global object', () => {
+    expect(() => stringify(global, { maxDepth: 10000 })).not.toThrow();
+  });
+};
+
+describe('Source', () => {
+  tests(src);
+});
+
+describe('Dist', () => {
+  tests(dist);
+});
